+++ package.json
@@ -24,6 +24,7 @@
     "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.10.5",
     "formik": "^2.2.9",
+    "moment": "^2.29.4",
     "notiflix": "^3.2.6",
     "react": "^18.1.0",
     "react-aria": "^3.24.0",
