+++ src/components/ChoosedMonth/useCalendar/useCalendar.js
@@ -1,0 +1,1 @@
+export const useCalendar = () => {};
